:recycle: asEquipmentResponse 함수 이동
